integrate NewsFetchService
